requirements.txt
dill
ipyaggrid==0.2.1
ipywidgets==7.4.1
pytest
tqdm
attrs==19.1.0
loguru==0.5.0
Pygments
colorama
autopep8
ipython
pandas==0.25.1
astunparse
